+++ tests/tools.py
@@ -1,2 +1,3 @@
 """Test helper classes and functions."""
+
 MOCK_SESSION_ID = "F9F9EE8505F55838"